initialize django properly move consumer initialization to a separate file to minimize imports
