+++ src/sentry/runner/commands/run.py
@@ -9,6 +9,7 @@ import click
 
 from sentry.bgtasks.api import managed_bgtasks
 from sentry.ingest.types import ConsumerType
+from sentry.issues.run import get_occurrences_ingest_consumer
 from sentry.runner.decorators import configuration, log_options
 from sentry.sentry_metrics.consumers.indexer.slicing_router import get_slicing_router
 from sentry.utils.kafka import run_processor_with_signals
@@ -534,7 +535,12 @@ def ingest_consumer(consumer_types, all_consumer_types, **options):
 
 
 @run.command("occurrences-ingest-consumer")
-@kafka_options("occurrence-consumer", include_batching_options=True, allow_force_cluster=False)
+@kafka_options(
+    "occurrence-consumer",
+    include_batching_options=True,
+    allow_force_cluster=False,
+    default_max_batch_size=100,
+)
 @strict_offset_reset_option()
 @configuration
 @click.option(
@@ -547,7 +553,6 @@ def ingest_consumer(consumer_types, all_consumer_types, **options):
 def occurrences_ingest_consumer(**options):
     from django.conf import settings
 
-    from sentry.issues.occurrence_consumer import get_occurrences_ingest_consumer
     from sentry.utils import metrics
 
     consumer_type = settings.KAFKA_INGEST_OCCURRENCES
